+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Place thumbs according to position on `ProductImages`.
+
 ## [3.21.8] - 2019-04-03
 
 ### Fixed